requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
jinja2>=3.1.3
python-multipart>=0.0.6
python-docx>=1.1.0
markdown>=3.5.2
httpx>=0.26.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
